+++ backend/canisters/user_index/api/src/updates/cancel_user_migration.rs
@@ -1,0 +1,16 @@
+use oc_error_codes::OCError;
+use serde::{Deserialize, Serialize};
+use types::{CanisterId, UserId};
+
+// Only available in test mode, until the UserIndex migrates users itself
+#[derive(Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub user_id: UserId,
+    pub multi_user_canister_id: CanisterId,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success,
+    Error(OCError),
+}